Created SpriteBatch.Native for native side `DynamicSpriteFont` draw commands
